+++ MDMI-F2014/HDF5Reader/StringAttribute.cs
@@ -6,7 +6,7 @@ namespace HDF5Reader
     class StringAttribute : Attribute
     {
 
-        public StringAttribute(string name, int length) : base(name,length) { }
+        public StringAttribute(string name, long length) : base(name,length) { }
 
         public override object Parse(byte[] data)
         {
